renamed glue_mrpc->glue and changes
